+++ src/types/camera.c
@@ -7,6 +7,7 @@
 #include "trace.h"
 #include "system/tls.h"
 #include "pool/pool.h"
+#include "renderers/render.h"
 
 #define GLHCK_CHANNEL GLHCK_CHANNEL_CAMERA
 
@@ -31,7 +32,7 @@ enum pool {
    POOL_LAST
 };
 
-static unsigned int pool_sizes[POOL_LAST] = {
+static size_t pool_sizes[POOL_LAST] = {
    sizeof(glhckHandle), // view
    sizeof(glhckFrustum), // frustum
    sizeof(glhckRect), // viewport
