+++ src/types/framebuffer.c
@@ -17,7 +17,7 @@ enum pool {
    POOL_LAST
 };
 
-static unsigned int pool_sizes[POOL_LAST] = {
+static size_t pool_sizes[POOL_LAST] = {
    sizeof(glhckRect), // rect
    sizeof(unsigned int), // object
    sizeof(glhckFramebufferTarget), // target
