+++ src/types/renderbuffer.c
@@ -19,7 +19,7 @@ enum pool {
    POOL_LAST
 };
 
-static unsigned int pool_sizes[POOL_LAST] = {
+static size_t pool_sizes[POOL_LAST] = {
    sizeof(struct dimensions), // rect
    sizeof(unsigned int), // object
    sizeof(glhckTextureFormat), // target
