+++ src/types/skeletal/animator.c
@@ -25,7 +25,7 @@ enum pool {
    POOL_LAST
 };
 
-static unsigned int pool_sizes[POOL_LAST] = {
+static size_t pool_sizes[POOL_LAST] = {
    sizeof(glhckHandle), // animation
    sizeof(glhckHandle), // states
    sizeof(glhckHandle), // bones
