+++ src/types/skeletal/node.c
@@ -16,7 +16,7 @@ enum pool {
    POOL_LAST
 };
 
-static unsigned int pool_sizes[POOL_LAST] = {
+static size_t pool_sizes[POOL_LAST] = {
    sizeof(glhckHandle), // rotations
    sizeof(glhckHandle), // translations
    sizeof(glhckHandle), // sclaings
