+++ src/types/skeletal/skinbone.c
@@ -26,7 +26,7 @@ enum pool {
    POOL_LAST
 };
 
-static unsigned int pool_sizes[POOL_LAST] = {
+static size_t pool_sizes[POOL_LAST] = {
    sizeof(kmMat4), // offsetMatrix
    sizeof(glhckHandle), // weights
    sizeof(glhckHandle), // bone
